Linear Emulator (#55)
* Sketch out ecall for reseting stack pointer.

* Start to set everything up for Harvard emulator.

* Add permissions to fixed-memory.

* Add permissions to variable memory.

* Add unified memory.

* Formatting

* Work on adding tests

* Get tests passing

* Most of way to integrating syscalls and private input tape.

* Get rest of tests passing.

* Format.

* Add layout struct.

* Add notion of unified memory without a variable segment.

* Fix unified tests.

* Add notion of no-access memory.

* Much more work on emulator trait.

* Format.

* Inline sp ecall.

* Rethink lifetimes...

* Add executor struct and emulator traits.

* Formatting.

* More fixing up.

* Formatting.

* Get most tests passing.

* Formatting.

* Debugging work on stack.

* Add heap overwriting as well.

* Formatting.

* Reorder checking for heap overwrite.

* More formatting.

* Rebase.

* Minor fixes.

* Fix confusion between physical and virtual addressing.

* Fix parsing.

* Renaming.

* Fix linker script + usage in integration tests.

* Move unit test binary.

* Format.

* Fix layout and get all tests passing.

* Format.

* More format.

* Remove debug

* Update vm/src/emulator/executor.rs


* Update vm/src/emulator/executor.rs


* Update vm/src/emulator/executor.rs


* Update vm/src/emulator/layout.rs


* Layout fixes and more efficient loading.

* Fix doctests.

* Clippy

* Format.

* Fixup.

* Resolve comments.

---------

+++ vm/src/riscv/instructions/macros.rs
@@ -72,7 +72,7 @@ macro_rules! impl_u_type_instructions {
                     $opcode,
                     dec_insn.rd as _,
                     0,
-                    (dec_insn.imm >> 12) as _,
+                    ((dec_insn.imm as u32) >> 12) as _,
                     InstructionType::UType,
                 )
             }
